pyproject.toml
[project]
name = "news-linkedin-publisher"
version = "0.1.0"
description = "Scraper de noticias y publicado en LinkedIn para AI, Data Science, ML, DL"
authors = [
    {name = "Jose Martin Rangel Espinoza",email = "jrangelespinoza@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "openai (==0.28.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "async-timeout (>=4.0.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"
